renamed database provider projects to Nox.EntityFramework.*
