+++ src/Nox.EntityFramework.Sqlite/Nox.EntityFramework.Sqlite.csproj
