+++ src/Nox.Generator/Nox.Generator.csproj
@@ -26,6 +26,7 @@
   <ItemGroup>
 	  <PackageReference Include="YamlDotNet" Version="13.1.1" PrivateAssets="all" GeneratePathProperty="true" />
     <PackageReference Include="Humanizer.Core" Version="2.14.1" PrivateAssets="all" GeneratePathProperty="true" />
+    <PackageReference Include="FluentValidation" Version="11.6.0" PrivateAssets="all" GeneratePathProperty="true" />
   </ItemGroup>
  
   <ItemGroup>
@@ -40,6 +41,7 @@
     <ItemGroup>
       <TargetPathWithTargetPlatformMoniker Include="$(PKGYamlDotNet)\lib\netstandard2.0\YamlDotNet.dll" IncludeRuntimeDependency="false" />
       <TargetPathWithTargetPlatformMoniker Include="$(PKGHumanizer_Core)\lib\netstandard2.0\Humanizer.dll" IncludeRuntimeDependency="false" />
+      <TargetPathWithTargetPlatformMoniker Include="$(PKGFluentValidation)\lib\netstandard2.0\FluentValidation.dll" IncludeRuntimeDependency="false" />
       <TargetPathWithTargetPlatformMoniker Include="..\..\src\Nox.Solution\bin\$(Configuration)\netstandard2.0\Nox.Solution.dll" IncludeRuntimeDependency="false" />
       <TargetPathWithTargetPlatformMoniker Include="..\..\src\Nox.Types\bin\$(Configuration)\netstandard2.0\Nox.Types.dll" IncludeRuntimeDependency="false" />
     </ItemGroup>
