+++ src/Nox.Lib/Nox.Lib.csproj
@@ -36,9 +36,9 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<ProjectReference Include="..\Nox.DatabaseProvider.Postgres\Nox.DatabaseProvider.Postgres.csproj" />
-		<ProjectReference Include="..\Nox.DatabaseProvider.Sqlite\Nox.DatabaseProvider.Sqlite.csproj" />
-		<ProjectReference Include="..\Nox.DatabaseProvider.SqlServer\Nox.DatabaseProvider.SqlServer.csproj" />
+		<ProjectReference Include="..\Nox.EntityFramework.Postgres\Nox.EntityFramework.Postgres.csproj" />
+		<ProjectReference Include="..\Nox.EntityFramework.Sqlite\Nox.EntityFramework.Sqlite.csproj" />
+		<ProjectReference Include="..\Nox.EntityFramework.SqlServer\Nox.EntityFramework.SqlServer.csproj" />
 		<ProjectReference Include="..\Nox.Solution\Nox.Solution.csproj" />
 		<ProjectReference Include="..\Nox.Abstractions\Nox.Abstractions.csproj" />
 		<ProjectReference Include="..\Nox.Logging\Nox.Logging.csproj" />
